assets: plug remove from community action to callbacks
