+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/landing_page/index.js
@@ -88,7 +88,6 @@ if (sidebarCommunitiesManageDiv) {
       recordCommunitySearchEndpoint={recordCommunitySearchEndpoint}
       permissions={JSON.parse(recordManagementAppDiv.dataset.permissions)}
       searchConfig={JSON.parse(pendingCommunitiesSearchConfig)}
-      record={JSON.parse(sidebarCommunitiesManageDiv.dataset.record)}
     />,
     sidebarCommunitiesManageDiv
   );
