+++ run-js-linter.sh
@@ -19,6 +19,9 @@ for arg in $@; do
 	case ${arg} in
 		-i|--install)
 			npm install --no-save --no-package-lock @inveniosoftware/eslint-config-invenio@^2.0.0;;
+	  -f|--fix)
+	    printf "${GREEN}Run eslint${NC}\n";
+      npx eslint -c .eslintrc.yml invenio_app_rdm/**/*.js --fix;;
 		*)
 			printf "Argument ${RED}$arg${NC} not supported\n"
 			exit;;
